Fix server action export boundaries
