+++ apps/web/app/(landing)/components/test-action.ts
@@ -5,6 +5,7 @@ import { sleep } from "@/utils/sleep";
 
 const logger = createScopedLogger("testAction");
 
+// server-action-export: allow - intentionally exposes a non-mutating test action.
 export async function testAction() {
   logger.info("testAction started");
