+++ apps/web/package.json
@@ -11,6 +11,7 @@
     "start:standalone": "node .next/standalone/server.js",
     "lint": "biome check .",
     "check-enums": "node scripts/check-enum-imports.js",
+    "check-server-actions": "node scripts/check-server-action-exports.js",
     "generate-llm-pricing": "tsx scripts/generate-llm-pricing.ts",
     "test": "cross-env RUN_AI_TESTS=false vitest",
     "test-ai": "cross-env RUN_AI_TESTS=true vitest --run",
