+++ turbo.json
@@ -317,6 +317,7 @@
     },
     "lint": {},
     "check-enums": {},
+    "check-server-actions": {},
     "dev": {
       "cache": false,
       "persistent": true
